estrip: obviate a race condition causing duplicate warnings
The estrip utility forks itself so as to achieve a degree of
parallelism. Unfortunately, it is also predisposed to raising warnings
deep into the function call stack. Although a given worker process makes
a point of never issuing the same type of warning more than once, it
remains unaware of the activities of another. This gives rise to a race
condition in which ewarn() may be invoked in parallel, causing redundant
warnings whose lines are unpredictably interleaved.

* FEATURES=installsources is enabled but the debugedit binary could not be
* FEATURES=installsources is enabled but the debugedit binary could not be
* found. This feature will not work unless debugedit is installed!
* found. This feature will not work unless debugedit is installed!

Work around this issue by writing out the warnings to files whose names
differ only by incorporating the checksum of the warning. In order to
obviate the race condition, a temporary file is written to, after which
mv(1) is used to atomically replace the existing file, if any. These
warnings are later collected and conveyed through the invocation of the
ewarn() function, just before estrip exits.

By no means do I consider this to be an ideal solution but it is, at
least, both a coherent and effective one; not to mention that it can
potentially be improved upon later.

Bug: https://bugs.gentoo.org/832138
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>