90gcc-warnings: de-deduplicate non-adjacent warnings
Presently, the 90gcc-warnings QA check uses uniq(1) to de-duplicate the
collected warnings. In doing so, it requires for duplicate lines to be
adjacent to one another in order to be effective. However, this will not
necessarily be the case. Instead, use awk(1) to de-duplicate the
warnings while preserving their original order.

Bug: https://bugs.gentoo.org/832138
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>