phase-functions.sh: reduce the degree of indentation in __dyn_install()
Presently, the __dyn_install() function guards the routine that records
the sizes of WORKDIR and D with a test for the existence of the du(1)
utility. Decrease the indentation level of the entire routine by
repositioning the test and simply exiting the subshell in the highly
unlikely event that the utility be missing. It is unclear to me why the
test is present, given that it is a standard utility and is required by
POSIX to exist.

Further, use the hash builtin instead of the type builtin. While type -P
does coerce bash into performing a PATH search, the manner in which du
is invoked still permits for it to be a function or alias. Using hash
also confers the benefit of pre-warming the cache that the shell employs
for Command Search and Execution.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>