misc-functions.sh: split words safely in {die,success}_hooks()
Presently, the die_hooks() and success_hooks() functions split the
'EBUILD_DEATH_HOOKS' and 'EBUILD_SUCCESS_HOOKS' variables by way of an
unquoted expansion, making them sensitive to the prevailing value of
IFS, as well as potentially performing pathname expansion on the
resulting words.

Address the issue by using read to separate the words into an array, and
by ensuring that all expansions are quoted.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>