bashrc-functions.sh: have register_{die,success}_hook() validate their parameters
Presently, the register_die_hook() and register_success_hook() functions
make use of the contains_word() function in order to determine whether
the 'EBUILD_DEATH_HOOKS' and 'EBUILD_SUCCESS_HOOKS' variables need to be
extended. This is so as to ensure that the same hook cannot be added
more than once.

  # Adds a hook for the my_cleanup function.
  register_die_hook my_cleanup

  # This is a no-op because the hook is already registered.
  register_die_hook my_cleanup

However, the contains_word() function will always return false where its
first argument contains whitespace, for which there is a good reason.

  # We may consider this to be a scalar variable comprising four words.
  wordlist='foo bar baz quux'

  # Thus, it would not be appropriate for this to return true.
  contains_word 'bar baz' "${wordlist}"

So far, so good. However, one must then consider that, whenever
contains_word() returns false, it is effectively granting permission for
the hook to be added again, giving rise to the following edge case.

  # Two hooks will be registered by the names of "bar" and "baz".
  # Not an intended use case; two arguments should be given.
  register_die_hook 'bar baz'

  # This will register both again, despite having been added already.
  register_die_hook 'bar baz'

To remedy this, have the register_die_hook() and register_success_hook()
functions validate their parameters. The nature of the validation is
straightforward: if a given parameter be empty, or contain any
whitespace, then it shall be considered as invalid and disregarded.

Though such cases should be vanishingly rare in practice, I would have
liked to raise a warning for them. However, /etc/portage/bashrc is
sourced so often that, barring an improved warning mechanism, it is
currently impractical.

Fixes: 3b0150e488972b02c45c71192507300d823a361f
See-also: 4a4631eef7186c29668a8c049d988b61469940fd
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>