ecompress: have guess_suffix() use a RETURN trap to purge tmpdir
Now that the guess_suffix() function refrains from enabling the errexit
option, it is possible to reliably employ a RETURN trap. Do so, thereby
guaranteeing the destruction of the temporary directory upon returning.

Also, localise the 'tmpdir' variable and refrain from needlessly
switching to the / directory. That the current working directory changes
is unimportant because the function is only ever called from a subshell.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>