ecompress: safely split the PORTAGE_COMPRESS{,FLAGS} variables
Presently, the guest_suffix() function attempts to execute a simple
command through the unquoted expansion of both the 'PORTAGE_COMPRESS'
and 'PORTAGE_COMPRESS_FLAGS' variables. This renders it sensitive to the
prevailing value of IFS, and runs the risk of performing pathname
expansion on the resulting words.

Address the issue by using read to safely separate the collected words
into an array, and by expanding the array as a simple command.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>