ecompress: improve the error handling that occurs before exiting
Refrain from invoking the fix_symlinks() function in the case that not
all of the files were successfully compressed.

In the case that not all the files were successfully compressed, have
the exception message clearly explain so.

In the case that not all of the broken symlinks were successfully
repaired, have the exception message clearly explain so.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>