ecompress: refine the error handling of fix_symlinks()
If either rm(1) or ln(1) fails, refrain from processing any outstanding
symlinks. Instead, return false immediately.

If find(1) fails, refrain from invoking die(). Instead, have the parent
shell wait for the enclosing process substitution to conclude before
returning false. To accomplish this, the wait builtin is invoked in a
manner that is acceptable as of bash >=4.4.

Drop the 'ret' variable, for it is no longer required.

It should be noted that the responsibility of throwing an exception is
already assumed by the calling context. That is, if fix_symlinks()
returns false for any reason, an exception shall be thrown.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>