bashrc-functions.sh: accelerate __strip_duplicate_slashes() by not looping
Accelerate the __strip_duplicate_slashes() function by using the
combination of a string-replacing parameter expansion and an extended
pattern matching operator to strip the slashes in a single pass.

Doing so depends upon the extglob option, which must be handled with a
commensurate degree of caution. It is important to understand that the
extglob option affects the behaviour of the parser. Thus, it is possible
to write code which is syntantically valid in view of the extglob option
being enabled, yet which is invalid otherwise.

$ bash -O extglob -c ': +(.)'         # ok
$ bash -c ': +(.)'                    # error
bash: -c: line 1: syntax error near unexpected token `('
$ bash -c 'shopt -s extglob; : +(.)'  # still an error!
bash: -c: line 1: syntax error near unexpected token `('
$ bash -c $'shopt -s extglob\n: +(.)' # ok

In the third case, an error occurs because bash encounters the pattern
matching operator before it has had an opportunity to enable the extglob
option. The fourth cases addresses this by concluding the invocation of
the shopt builtin with a <newline>. Doing so gives bash the opportunity
to enable the option before the command bearing the pattern is parsed.

Consider, also, the following sample program.

#!/bin/bash
lowernames() {
    # Enable the extglob option upon the function being called.
    shopt -s extglob
    printf '%s\n' +([a-z])
}
lowernames

Bash will fail to parse the program as a whole, simply because the
extglob option is not enabled at the point that the +([a-z]) pattern is
seen and (unsuccessfully) parsed.

Now, since portage does not enable extglob of its own volition, it
presents a challenge. How does one prevent bash from being offended by
the presence of an extended pattern before the opportunity has arisen to
enable the extglob option at runtime? As it happens, the answer is a
remarkably simple one, which is to use eval to execute a string
containing the otherwise offending pattern.

#!/bin/bash
lowernames() {
   # Enable the extglob option upon the function being called.
   shopt -s extglob
   # Conceal the pattern from bash during the parsing stage. This works
   # because eval operands are never evaluated until eval is executed.
   eval 'printf "%s\n" +([a-z])'
}
lowernames

The revised __strip_duplicate_slashes() function makes use of eval for
exactly this reason. Apart from evaluating the 'reset_extglob' payload,
there is no property of eval being exploited other than its ability to
conceal code from bash during the parsing stage.

Finally, use the printf builtin instead of echo.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>