ecompress: accelerate fix_symlinks() by having find(1) resolve targets
The GNU implementation of the find(1) utility offers some powerful
primaries. Enhance the performance of the fix_symlinks() function by
taking advantage of them. The exact manner in which this is accomplished
is described herewith.

Use the -xtype l primary so as not to waste any time processing
non-dangling symlinks. The shell is duly relieved of the need to test
for the existence of the link's target.

Use the -printf primary so as to print NUL-terminated pairs, where the
first item is the path of the symlink, and the second item its target.
Not only does this avoid an expensive invocation of readlink(1) for each
iteration of the loop but it also guarantees that any possible target
will be processed correctly. Hitherto, this had not been the case, for
the following reasons.

- the <newline> character can legally be present within a pathname
- readlink(1) arbitrarily produces a trailing <newline> character
- command substitutions strip all trailing <newline> characters

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>