ecompress: refrain from calling die() inside process substitutions
I care not for the approach of invoking the die() function from inside a
process substitution. Given that they are asynchronous in nature, one
ought to be mindful of the prospect of a race condition occurring.

Instead, have the parent shell wait for process substitutions to
conclude before collecting their exit status values and choosing whether
to call die(). To accomplish this, the wait builtin is invoked in a
manner that is acceptable as of bash >=4.4.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>