ecompress: don't process pathnames containing <newline>
In some instances, ecompress produces and consumes streams consisting of
null-terminated pathnames. In other instances, ecompress produces and
consumes streams consisting of <newline>-terminated pathnames. In
particular, the temporary files comprise <newline>-terminated pathnames.
This requirement is further entrenched by the manner in which the
do_ignore() function makes use of the sort(1) and comm(1) utilities.

In other words, there are two incompatible data models at play. Thus, it
is imperative that the the constraints imposed by the lesser model are
adhered to. The only way to do that is to refuse to process pathnames
that contain the <newline> in any capacity. Impose this necessary
constraint by introducing the changes described herewith.

Have both the do_ignore() and do_queue() functions disregard operands
that contain <newline>. This concerns the composition of the
".ecompress_skip_files" and ".ecompress_had_precompressed" files,
respectively.

Have the do_queue() function invoke find(1) in such a way that any
pathname containing <newline> is disregarded. This concerns the
composition of the ".ecompress_had_precompressed" file.

In all cases where find(1) is charged with locating files whose names
match the "*.ecompress" glob, invoke it in such a way that any pathname
containing <newline> is disregarded.

I would add that, while these changes will almost certainly be invisible
in nature - and will remain largely unappreciated - there can be no
excuse for dodging the issue.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>