helper-functions.sh: ensure __multijob_finish() returns numerically
In the course of making a minor change to the __multijob_finish()
function - intended to idiomatically employ the (( … )) command - the
initial assignment of 0 to the 'ret' variable was inadvertently removed.
Since it may not otherwise be assigned, the following error may arise.

helper-functions.sh: line 64: return: : numeric argument

Restore the assignment.

Fixes: 0581fbc0421e88d3ef01b55ab864803070be3d96
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>