isolated-functions.sh: have find0() support the {H,L,P} opts in backward-compat mode
GNU find(1) supports several options, the most important of which are
-H, -L and -P. The find0() function presently implements a fallback
mode, the purpose of which is to retain compatibility with the outdated
GitHub CI environment (which has an older version of findutils).

However, the fallback mode is currently unable to handle the
aforementioned options correctly. Render it able to do so by recognising
where they are present and ensuring that they precede both pathnames and
primaries alike. This may prove useful in the future.

See-also: a0210b3c49ea346b1e999f92a7ed89802e8d6849
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>