ecompress: rename several variables in do_queue()
Rename the 'collisions' variable to 'collision_by' so as to emphasise
that it is a map and not a list. Also, have its values be the null
string rather than "1", thereby saving a little memory.

Rename the 'vpath' variable to 'uncompressed_path', and the 'comp'
variable to 'suffix'. These new names better convey their respective
purposes.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>