ecompress: use mktemp(1) in guess_suffix()
Though mktemp(1) is not a standard utility, it is already being employed
elsewhere. Namely by "emerge-webrsync", "estrip", "etc-update" and the
"helper-functions.sh" library. Should the decision be made in the future
that it is inappropriate to rely on the presence of the utility, there
remains the possibility of implementing it as a function in the case
that the utility is found to be unavailable.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>