ecompress: sort colliding pathnames in QA warning
Where the do_queue() function determines that there are any colliding
pathnames, lexicographically sort them at the point of conveying them
with eqawarn().

Further, employ the ${param@Q} form of expansion so as to ensure that
all pathnames are legible, and that control characters and (some)
whitespace characters are not printed in their raw form. Doing so is
acceptable, given a target of >=bash-4.4.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>