ecompress: always treat PORTAGE_COMPRESS as a discrete command name
Recently, the guess_suffix() function was made to use the read builtin
to safely split some of the words contained by the environment variables
'PORTAGE_COMPRESS' and 'PORTAGE_COMPRESS_FLAGS', like so.

  read -rd '' -a cmd <<<"${PORTAGE_COMPRESS} ${PORTAGE_COMPRESS_FLAGS}"

However, to split the value of the former variable is improper. The
status quo is that its value always be treated as the command name,
meaning that it should not be split. Address this by splitting only the
value of 'PORTAGE_COMPRESS_FLAGS'.

Note that I am not designating this as a fix for my prior commit because
the prior code was also incorrect in this respect. It was only upon
studying ecompress-file - which properly quotes the expansion of
'PORTAGE_COMPRESS' - that this long-standing discrepancy became clear.

See-also: a722142b8676a46c432e77133ec97d802edd1845
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>