isolated-functions.sh: don't run find(1) where the paths var is empty
Presently, the find0() function implements a fallback mode to cover
situations where the -files0-from primary is unavailable, thereby
retaining compatibility with the GitHub CI environment (which suffers
from an outdated version of findutils).

Consider the case where find0() is given no input. It will effectively
do nothing, successfully.

$ : | find0 -print; echo "$?"
0

However, if find0() determines that it needs to employ the fallback
mode, it may instead act as if "." is to be searched.

$ cd /etc/skel
$ : | find0 -print
.
./.bash_logout
./.bash_profile
./.bashrc

Though it is rather unlikely that anyone will ever commit code to
portage that is capable of exposing this discrepancy, I consider the
matter to be worth addressing. Do so by refraining from executing
find(1) in the case that the 'paths' array variable is empty.

See-also: a0210b3c49ea346b1e999f92a7ed89802e8d6849
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>