estrip: refactor some of the code out into functions
Relocate the code that attends to the --ignore and --queue options into
two new functions, respectively named do_ignore() and do_queue().

It should be noted that the code contained by the new functions is
over-indented so as to prevent the patch from being incomprehensible.
The indentation shall be adjusted by the next commit.

See-also: d015284bbdcdd36b5932a882d11fef9bf0941b96
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>