estrip: marginally improve the handling of STRIP_MASK
The 'STRIP_MASK' variable is intended to contain a <blank>-separated
series of zero-or-more words, each of which is to be treated as an
extended globbing pattern. Should a given pathname match any of these
patterns, it must not be stripped. This commit improves the manner in
which the variable is handled, as described herewith.

Rename the 'm' variable to 'glob', for clarity.

Refrain from using echo to print the words produced by expanding the
'STRIP_MASK' variable. It suffers from various portability issues, and
may respond to "special first operands", as Ramey puts it. Instead, use
the printf builtin.

Consume the words produced by expanding the 'STRIP_MASK' variable as a
stream of zero-or-more null-terminated records. This makes it possible
to employ certain patterns that could not have worked before. For
instance, consider the following assignment (using shell syntax rather
than make.conf syntax).

STRIP_MASK="/foo /bar/\$'\t'baz"

Prior to this commit, the value would map to these three globs:

declare -- glob="/foo"
declare -- glob="/bar/"
declare -- glob="baz"

Following this commit, the value maps to these two globs:

declare -- glob="/foo"
declare -- glob=$'/bar/\tbaz'

Unfortunately, the invocation of the eval builtin remains, since it is
required to support brace expansion. I have added an excoriating comment
regarding this because it is an absurd anti-feature that ought to be
dropped. Until such time as that happens, arbitrary code execution will
remain an attendant risk.

Consider a plausible scenario, in which a user might find it convenient
to define, say, "/foo/{bar,baz,quux}" as a mask. Instead, it should be
written as "/foo/@(bar|baz|quux)", which is a perfectly valid extglob.

Finally, suppress an instance of SC2053 that is a false-positive.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>