estrip: don't treat find_path as an ERE sub-pattern
Presently, the do_queue() function performs the following test in the
course of consuming pathnames from the scanelf(1) utility.

[[ "${needed_entry_file}" =~ ^${find_path##"${D}"} ]]

Given that the expansion of 'find_path' is unquoted, its value is
treated as a component of an extended regular expression. Address this
by properly quoting the expansion.

Use the == operator to match an equivalent glob, as opposed to the =~
operator. ERE matching is always slower in bash.

Refrain from employing the ${param##"$prefix"} form of expansion, given
that ${param#"$prefix"} does exactly the same thing while properly
communicating the intent.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>