Use mktemp(1) in a consistent and dependable fashion
Though mktemp(1) is not a standard utility in the de jure sense, it is
already being employed by the following units.

- bin/ebuild-helpers/doexe
- bin/ebuild-helpers/newins
- bin/ecompress
- bin/emerge-webrsync
- bin/estrip
- bin/etc-update
- bin/helper-functions.sh
- misc/emerge-delta-webrsync

The motivation for this commit is to use mktemp(1) in a consistent way,
known to be compatible with the majority of popular implementations and
resulting in broadly predictable behaviour. Such is accomplished by
adhering to the following set of recommendations.

- DO specify the -d or -u options, where appropriate
- DO specify the template in full, including the intended parent directory
- DO use -- to separate the template operand from the options
- DON'T specify any options other than -d or -u (for reasons of portability)
- DON'T specify more than six <X> chars in the template (for busybox support)
- DON'T assume TMPDIR will influence mktemp(1) on account of being an env var

Accordingly, this commit adjusts all eight of the aforementioned units.

In many cases, the parent directory of the temporary file is the value
of the 'T' variable, the expansion of which has been adjusted so as to
permit "/tmp" as a fallback value. The reason for this is to gradually
render various components of portage easier to test in isolation.

Some of the units merit additional commentary, which is as below.

# bin/ebuild-helpers/doexe

Define an EXIT trap to guarantee the subsequent destruction of the
temporary directory.

Check the exit status of mktemp(1).

Assign the output of mktemp(1) to a variable whose name is in lowercase,
thereby reducing the probability of a namespace conflict and clarifying
that the variable is not expected to be defined by the parent process.

# bin/ebuild-helpers/newins

Assign the output of mktemp(1) to a variable whose name is in lowercase

Set up the EXIT trap safely by ensuring that the 'tmpdir' variable is
assigned the null string beforehand (rather than trust the environment).

# bin/emerge-webrsync

Assign the output of mktemp(1) to a variable whose name is in lowercase.
To name the variable 'TMPDIR' was asburd, since its value can influence
the behaviour of other programs (if exportable).

Define the EXIT trap safely by ensuring that the 'tmpdir' variable is
assigned the null string beforehand.

Have the trap execute neither of the "set -u" nor "cd /" commands, for
both are unnecessary.

Don't issue needless test -w commands after executing mktemp(1). Rather,
check whether mktemp(1) succeeds.

# bin/estrip

Define an EXIT trap to guarantee the subsequent destruction of the
temporary directory.

# bin/etc-update

Define an EXIT trap to guarantee the subsequent destruction of the
temporary directory.

Have the show_diff() function define a RETURN trap to guarantee the
subsequent destruction of its localised temporary directory.

# bin/helper-functions.sh

Specify the -u option, thereby preventing mktemp(1) from creating an
unwanted regular file.

# misc/emerge-delta-webrsync

Assign the output of mktemp(1) to a variable whose name is in lowercase.
Again, 'TMPDIR' was an absurd choice.

Define the EXIT trap safely by guaranteeing that the 'tmpdir' variable
is assigned the null string beforehand.

Don't issue needless test -w commands after executing mktemp(1). Rather,
check whether mktemp(1) succeeds.

See-also: e32009dfab8f4d22e3fb29817cd493ef1ff40768
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>