emerge-webrsync: improve PGP verification method selection in handle_pgp_setup()
Presently, the emerge-webrsync program contains the handle_pgp_setup()
function. Its duty is to determine whether the user intends for GPG
verification to be employed. I would describe it as being in a state of
disrepair, for the following reasons.

- the logic is not readily apparent at a glance
- the comments have not kept pace with the code
- the "sync-webrsync-verify-signature" repo attribute test is broken and defunct
- the test for whether the program was launched by the webrsync module is broken
- a user-hostile exception is issued in one particular instance

This commit addresses these issues in the ways described herewith.

Ensure that the --no-pgp-primacy option is granted primacy by testing
for it first. Not only is this behavior conventional but it also allows
for the broken "sync-webrsync-verify-signature" attribute handling code
to be dropped outright. As of commit b39f9f8, the webrsync module makes
a point of passing the option to emerge-webrsync where appropriate.

In the case that "webrsync-gpg" is found to be present in FEATURES,
refrain from incorrectly testing whether emerge-webrsync was invoked
'directly' by the user and, if so concluded, throwing a user-hostile
exception. I had discerned a more reliable means of performing this
test and considered whether the diagnostic might be improved. However,
following a discussion with Sam, it was concluded that it would be
simpler to drop the test. Consequently, a warning shall always be
issued but never an exception thrown.

For each branch that assigns a particular value to the
'WEBRSYNC_VERIFY_SIGNATURE' variable, ensure that the accompanying
comment is as concisely helpful as possible.

Remove a redundant pattern from a case statement on the basis that there
is no circumstance under which 'WEBRSYNC_VERIFY_SIGNATURE' will have a
value that cannot match [012] at that juncture.

See-also: b39f9f819c34a7a67f2639acb4d4c17a1aa5df89
Bug: https://bugs.gentoo.org/940120
Bug: https://bugs.gentoo.org/945861
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>