emerge-webrsync: perform minor refactoring upon check_file_signature()
This commit performs some minor refactoring upon the
check_file_signature() function, the exact nature of which is described
herewith.

Drop the 'if' command that is responsible for testing the value of the
'WEBRSYNC_VERIFY_SIGNATURE' variable and potentially displaying an
informational message. This responsibility is now handled by the
previously enclosed case statement by matching against an additional
pattern and deliberately falling through to the next. The maximal degree
of identation is thus reduced.

Refrain from assigning 1 to the 'r' variable as a foregone conclusion,
and drop the superfluous 'return' command. Should verification fail, the
die() function is called. Otherwise, the function shall always
implicitly return 0.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>