emerge-webrsync: have check_file_signature() handle exit codes flexibly
Presently, the check_file_signature() function requires for the 'r'
variable to contain a meaningful exit status code to be able to
determine whether to invoke die(). Instead, employ the ${r=$?} form of
parameter expansion so that the value of the '?' parameter is
substituted and assigned for cases where 'r' hasn't been set.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>