emerge-webrsync: situate the && operator at the front for command chains
As concerns the do_snapshot() function, there are two instances in which
batches of commands are composed by joining them with the && control
operator, such that each is contingent upon the last succeeding. For
both of these instances, have the operators precede the commands rather
than follow them. While only a cosmetic change, it does help to clarify
the control flow.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>