emerge-webrsync: don't let check_file_signature_gpg_unwrapped() modify the r variable
Presently, the check_file_signature_gpg_unwrapped() sets the 'r'
variable to a value of 0 upon confirming that the PGP signature is good.
However, the 'r' variable is not its property; it actually belongs to
the check_file_signature() function in its capacity as a caller.

Rather than modify the variable in question, have the function return
successfully upon confirming that the signature is good.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>