emerge-webrsync: fix insecure key verification in check_file_signature_gpg_unwrapped()
Presently, the check_file_signature_gpg_unwrapped() function exists so
as to be able to support the legacy "webrsync-gpg" option, which may be
defined by FEATURES and is (barely) documented by make.conf(5). It is
also called upon in cases where gemato would be used but cannot be
found, for whatever reason.

Unfortunately, the way in which it performs signature verification is
insecure. It captures the standard output of gpg(1) and tries to match
any of its lines against this pattern:

  "[GNUPG:] GOODSIG"*

Firstly, it is incorrect in that it does not require for "GOODSIG" to be
followed by a space. Secondly, and rather more seriously, it is simply
insufficient to match the GOODSIG token in isolation. Consider the
following gpg(1) output sample, which has been reduced so as to show
only the GOODSIG, VALIDSIG and TRUST_ULTIMATE tokens.

[GNUPG:] GOODSIG EC590EEAC9189250 Gentoo ebuild repository signing key (Autom...
[GNUPG:] VALIDSIG E1D6ABB63BFCFB4BA02FDF1CEC590EEAC9189250 2025-06-15 1749949...
[GNUPG:] TRUST_ULTIMATE 0 pgp

In order to be sure that the key is ultimately trusted, that its
signature is valid, and that nothing has expired, it is necessary to
match against all three of the aforementioned tokens.

Consider, also, the following output, for which the key was at a trust
level of 4 (fully trusted), rather than 5 (ultimately trusted).

[GNUPG:] GOODSIG EC590EEAC9189250 Gentoo ebuild repository signing key (Autom...
[GNUPG:] VALIDSIG E1D6ABB63BFCFB4BA02FDF1CEC590EEAC9189250 2025-06-15 1749949...
[GNUPG:] TRUST_UNDEFINED 0 pgp

It demonstrates that there is little sense in trying to match for keys
whose trust levels are lower than 5. Indeed, it is possible for there to
be a sequence comprised by all three of GOODSIG, VALIDSIG and
TRUST_NEVER. Such would indicate that the signature is invalid, yet the
current code would consider it to be valid.

Address this defect by introducing a helper function named gpg_verify(),
which is responsible for executing gpg --verify and capturing its
output, then efficiently matching its output against all three of these
tokens: GOODSIG, VALIDSIG, TRUST_ULTIMATE.

[sam: Note that while this isn't great, it's also not as important
as it would otherwise be, given gemato is in stage3s and we use it
whenever available since the fixes for bug #597800. The code here is
only a fallback option and we warn when using it.]

Bug: https://bugs.gentoo.org/597800
Fixes: ffd68477e5c1e1badf60c86ae221c90dad50390d
Link: https://www.gnupg.org/documentation/manuals/gnupg/Automated-signature-checking.html
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>