emerge-webrsync: address a slew of shellcheck warnings
As concerns the fetch_file() function, suppress an instance of SC2034.
It is a false positive because the subsequent use of the 'URI' variable
is concealed by the use of the eval builtin.

As concerns the sync_local() function, address five instances of SC2086
by quoting the expansions of various variables, and by declaring and
using the 'chown_opts' and 'rsync_opts' variables as arrays.

A concerns the do_snapshot() function, address an instance of SC2004 by
refraining from performing an unnecessary parameter expansion upon the
'utc_seconds' variable in the arithmetic context.

A concerns the do_latest_snapshot() function, address an instance of
SC2004 by refraining from performing an unnecessary parameter expansion
upon the 'attempts' variable in the arithmetic context.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>