emerge-webrsync: clarify behaviour where gpg(1) is missing
Presently, the check_file_signature_gpg_unwrapped() function checks
whether gpg(1) is in PATH. In order to find out what happens in the case
that the check fails, it is necessary to read a comparatively long
branch of code. Remedy this by inverting the sense of the test and
making it clear that an exception shall be thrown where gpg(1) is found
to be missing.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>