emerge-webrsync: fix fallback mode where PORTAGE_GPG_DIR is unset/empty
Recently, the method by which the check_file_signature_gpg_unwrapped()
function verifies key signatures was found to be insecure. The issue was
addressed by requiring for the key to be "ultimately" trusted and by
ensuring that all three of the GOODSIG, VALIDSIG and TRUST_ULTIMATE
tokens are observable in the output stream of gpg(1). However, this
breaks the case where all three of the following conditions hold true.

1) the user hasn't defined FEATURES="webrsync-gpg", per make.conf(5)
2) the user hasn't set PORTAGE_GPG_DIR, or has set it as an empty string
3) gemato is not found in PATH (improbable, though possible)

Under such circumstances, emerge-webrsync will attempt to populate an
ephemeral keyring by importing the distributed public key block from
"/usr/share/openpgp-keys/gentoo-release.asc". The problem is that the
relevant key will have an undefined trust level (TRUST_UNDEFINED), which
cannot satisfy the newly imposed strictures.

Remedy this by identifying the fingerprint of the key with an email
adddress of <infrastructure@gentoo.org> and automatically raising its
trust level to TRUST_ULTIMATE (which has an internal numerical level of
6). That way, there is no need to curtail the strict verification
process and adversely affect other scenarios.

Bug: https://bugs.gentoo.org/597800
Fixes: c5247ab0ea0b717f4037b6aa14108d7e41b5e2dd
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>