Refrain from using RETURN traps
Presently, there are two units that set up a RETURN trap.

- bin/ecompress (by the guess_suffix() function)
- bin/etc-update (by the show_diff() function)

Alas, RETURN traps are difficult to manage in bash because they affect
all functions until otherwise cleared or re-defined - a fact that I had
somehow managed to overlook at the time of introducing them. Whatever
value they would otherwise have conferred over the use of an EXIT trap
is duly decimated.

As concerns the ecompress utility, address the issue by having the
guess_suffix() function declare an EXIT trap instead. To do so is
acceptable, because bash is always made to fork in the course of calling
that particular function. So as to emphasise this point, the body of the
function has been changed to a compound command of the ( … ) form.

As concerns the etc-update utility, simply remove the RETURN trap. The
directories created by the show_diff() function are subdirectories of a
known temporary directory, which is already covered by an EXIT trap.

Fixes: 9b95224283b2ed992c482956df952fada2fd41ef
Fixes: 508a79cedd09079174299239afb4e12e5470b129
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>