emerge-webrsync: render einfo() a no-op for the --quiet opt
It is apparent from reading the code that one of the intended effects of
the --quiet option is to slience informational messsages that would
otherwise be conveyed by the einfo() function. Rather than repeatedly
check whether the option was specified, redeclare said function as one
that does nothing, if needs be.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>