emerge-webrsync: simplify the do_tar() function
Simplify the do_tar() function by having the case statement assume
responsibility for dispatching the decompression command, and by having
it be the first stage of the pipeline. Further, replace the buggy
PIPESTATUS handling (sensitive to the value of IFS) by enabling the
pipefail shell option in function scope. Doing so is acceptable, given a
target of bash >=4.4.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>