emerge-webrsync: always pass --fail and --show-error to curl(1)
Presently, the fetch_file() function extends the value of 'FETCHCOMMAND'
with the -f (--fail) option, but only where both of the following
conditions hold true:

1) curl(1) has been detected as the user-agent to be executed
2) the user passed the --quiet option to emerge-webrsync

The second of these conditions is inappropriate because -f affects how
curl(1) chooses its exit status value. In particular, the option ensures
that it exits with a status other than zero in the event that the HTTP
server responds with a code between 400 and 499. This behaviour is
useful for scripting purposes and should not be contingent upon whether
the user has requested for emerge-webrsync to be quiet.

Address this issue by ensuring that curl(1) is given both the -f and -S
(--show-error) options, unconditionally. The latter option prevents the
-s option from suppressing all diagnostics in the event of an error.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>